extract CSV validator
[Single Responsibility Principle](https://en.wikipedia.org/wiki/Single-responsibility_principle)

Also, extract CsvParsingUtil to avoid duplicate code
