+++ src/main/java/org/example/csv2tex/csv/CsvToSchoolReportDataParser.java
@@ -5,16 +5,16 @@ import org.apache.commons.csv.CSVRecord;
 import org.apache.commons.lang3.tuple.Pair;
 import org.example.csv2tex.data.SchoolCompetencyData;
 import org.example.csv2tex.data.SchoolReportData;
-import org.example.csv2tex.exception.InvalidCsvException;
 
-import java.io.*;
+import java.io.File;
+import java.io.FileReader;
+import java.io.IOException;
+import java.io.Reader;
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
-import static java.util.stream.Collectors.joining;
-import static org.example.csv2tex.exception.InvalidCsvCause.*;
+import static org.example.csv2tex.csv.CsvParsingUtil.isLevelSettingColumn;
+import static org.example.csv2tex.csv.CsvParsingUtil.splitCompetencyColumnHeader;
 
 public class CsvToSchoolReportDataParser {
 
@@ -23,7 +23,7 @@ public class CsvToSchoolReportDataParser {
         List<String> headers = pair.getLeft();
         List<CSVRecord> recordList = pair.getRight();
 
-        ensureCorrectFormat(headers, recordList);
+        new CsvValidator().ensureCorrectFormat(headers, recordList);
 
         List<SchoolReportData> result = new ArrayList<>();
         for (CSVRecord rawData : recordList) {
@@ -49,153 +49,6 @@ public class CsvToSchoolReportDataParser {
         // skip but save first record - header
         List<String> headerRow = recordList.remove(0).toList();
         return Pair.of(headerRow, recordList);
-    }
-
-    private void ensureCorrectFormat(List<String> headers, List<CSVRecord> recordList) {
-        ensureLevelColumn(headers);
-        ensureRowLengths(headers, recordList);
-        ensureCompleteCompetencyHeaders(headers);
-    }
-
-    private void ensureLevelColumn(List<String> headers) {
-        int baseDataLength = 6;
-        boolean levelColumnFound = false;
-        for (String columnName : headers) {
-            if (isLevelSettingColumn(columnName)) {
-                levelColumnFound = true;
-                break;
-            }
-        }
-        if (headers.size() > baseDataLength && !levelColumnFound) {
-            throw new InvalidCsvException(HEADER_NO_LEVEL_DEFINED);
-        }
-    }
-
-    private void ensureRowLengths(List<String> headers, List<CSVRecord> recordList) {
-        if (recordList.isEmpty()) {
-            return;
-        }
-
-        int headerRowLength = headers.size();
-        Map<Integer, Integer> rowLengths = countRowLengths(recordList);
-
-        boolean lengthChecksPassed = checkForSystematicFaults(headerRowLength, rowLengths);
-        if (lengthChecksPassed) {
-            return;
-        }
-        checkFaultsWithHeuristics(headerRowLength, rowLengths, recordList);
-    }
-
-    // checks cases where all content rows differ from header row
-
-    private Map<Integer, Integer> countRowLengths(List<CSVRecord> recordList) {
-        Map<Integer, Integer> rowLengths = new HashMap<>();
-        recordList.forEach(record -> {
-            int rowLength = record.size();
-            Integer existingValue = rowLengths.getOrDefault(rowLength, 0);
-            rowLengths.put(rowLength, existingValue + 1);
-        });
-        return rowLengths;
-    }
-
-    private boolean checkForSystematicFaults(int headerRowLength, Map<Integer, Integer> rowLengths) {
-        if (rowLengths.size() == 1) {
-            Integer contentRowsLength = rowLengths.keySet().iterator().next();
-            if (headerRowLength == contentRowsLength) {
-                // all rows have the same number of columns
-                return true;
-            } else if (headerRowLength > contentRowsLength) {
-                throw new InvalidCsvException(HEADER_LONGER_THAN_ALL_CONTENT);
-            } else if (headerRowLength < contentRowsLength) {
-                throw new InvalidCsvException(HEADER_SHORTER_THAN_ALL_CONTENT);
-            }
-        }
-        return false;
-    }
-
-    /**
-     * Heuristic: determine whether to show "some data rows are wrong" or "the header row is wrong"
-     */
-    private void checkFaultsWithHeuristics(float headerRowLength, Map<Integer, Integer> rowLengths, List<CSVRecord> recordList) {
-        float contentRowsShorterThanHeader = 0f;
-        float contentRowsLongerThanHeader = 0f;
-        for (Map.Entry<Integer, Integer> entry : rowLengths.entrySet()) {
-            if (entry.getKey() > headerRowLength) {
-                contentRowsLongerThanHeader += entry.getValue();
-            } else if (entry.getKey() < headerRowLength) {
-                contentRowsShorterThanHeader += entry.getValue();
-            }
-        }
-
-        // more than 20% / at least 10 rows off --> assume header error
-        float ratioWrongLengthRows = (contentRowsLongerThanHeader + contentRowsShorterThanHeader) / recordList.size();
-        if (ratioWrongLengthRows > 0.20f || contentRowsShorterThanHeader + contentRowsLongerThanHeader > 9) {
-            if (contentRowsLongerThanHeader > contentRowsShorterThanHeader) {
-                throw new InvalidCsvException(HEADER_SHORTER_THAN_CONTENT, getListOfLongerRows(headerRowLength, recordList));
-            } else {
-                throw new InvalidCsvException(HEADER_LONGER_THAN_CONTENT, getListOfShorterRows(headerRowLength, recordList));
-            }
-        } else { // assume content error
-            if (contentRowsLongerThanHeader > contentRowsShorterThanHeader) {
-                throw new InvalidCsvException(CONTENT_ROW_LONGER_THAN_HEADER, getListOfLongerRows(headerRowLength, recordList));
-            } else {
-                throw new InvalidCsvException(CONTENT_ROW_SHORTER_THAN_HEADER, getListOfShorterRows(headerRowLength, recordList));
-            }
-        }
-    }
-
-    private String getListOfLongerRows(float headerRowLength, List<CSVRecord> recordList) {
-        StringBuilder sb = new StringBuilder("");
-        for (int i = 0; i < recordList.size(); i++) {
-            CSVRecord record = recordList.get(i);
-            if (isLongerThanHeader(headerRowLength, record)) {
-                if (sb.length() > 0) {
-                    sb.append(",");
-                }
-                sb.append(i + 1);
-            }
-        }
-        return sb.toString();
-    }
-
-    private boolean isLongerThanHeader(float headerRowLength, CSVRecord record) {
-        return record.size() > headerRowLength;
-    }
-
-    private String getListOfShorterRows(float headerRowLength, List<CSVRecord> recordList) {
-        StringBuilder sb = new StringBuilder("");
-        for (int i = 0; i < recordList.size(); i++) {
-            CSVRecord record = recordList.get(i);
-            if (isShorterThanHeader(headerRowLength, record)) {
-                if (sb.length() > 0) {
-                    sb.append(",");
-                }
-                sb.append(i + 1);
-            }
-        }
-        return sb.toString();
-    }
-
-    private boolean isShorterThanHeader(float headerRowLength, CSVRecord record) {
-        return record.size() < headerRowLength;
-    }
-
-
-    private void ensureCompleteCompetencyHeaders(List<String> headers) {
-        // check every column after the base data columns
-        List<Integer> invalidHeaderColumnsOneBased = new ArrayList<>();
-        for (int i = 6; i < headers.size(); i++) {
-            String header = headers.get(i);
-            if (!isLevelSettingColumn(header)) {
-                if (splitCompetencyColumnHeader(header).length < 2) {
-                    invalidHeaderColumnsOneBased.add(i + 1);
-                }
-            }
-        }
-        if (!invalidHeaderColumnsOneBased.isEmpty()) {
-            throw new InvalidCsvException(HEADER_COMPETENCY_INCOMPLETE_DEFINITION,
-                    invalidHeaderColumnsOneBased.stream().map(Object::toString).collect(joining(",")));
-        }
     }
 
     private SchoolReportData createReportDataFromRecord(List<String> headers, CSVRecord rawRowData) {
@@ -230,10 +83,6 @@ public class CsvToSchoolReportDataParser {
         }
     }
 
-    private boolean isLevelSettingColumn(String columnHeader) {
-        return "niveau".equalsIgnoreCase(columnHeader) || "level".equalsIgnoreCase(columnHeader);
-    }
-
     private SchoolCompetencyData addCompetencyData(String columnHeader, String currentLevel, String cellValue) {
         SchoolCompetencyData competencyData = new SchoolCompetencyData();
         String[] columnHeaderRows = splitCompetencyColumnHeader(columnHeader);
@@ -254,8 +103,4 @@ public class CsvToSchoolReportDataParser {
         return competencyData;
     }
 
-    private String[] splitCompetencyColumnHeader(String columnHeader) {
-        // limit 4: will keep all lines after the 4th as part of the 4th split
-        return columnHeader.split("[\r\n]+", 4);
-    }
 }
